requirements.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.3
django-colorfield==0.1.15
django-cors-headers==2.4.0
django-filter==2.0.0
django-js-asset==1.1.0
django-mptt==0.9.1
djangorestframework==3.9.0
djangorestframework-recursive==0.1.2
drf-tweaks==0.8.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
Pillow==5.3.0
pytz==2018.7
requests==2.20.1
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.24.1
